.claude/skills/strands-agents/assets/custom-tools/requirements.txt
strands-agents[bedrock]
strands-agents-tools
